feat: Removed all services from context
Only types remain - services moved to desktop app
