+++ src/index.ts
@@ -2,14 +2,5 @@ export * from './BlockTargetProvider';
 export * from './BlockTypeProvider';
 export * from './ResourceTypeProvider';
 
-export * from './services/ClusterConfig';
-export * from './services/AssetService';
-export * from './services/BlockService';
-export * from './services/FileSystemService';
-export * from './services/IdentityService';
-export * from './services/InstanceService';
-export * from './services/PlannerService';
-export * from './services/SocketService';
-export * from './services/TrafficService';
-export * from './services/TaskService';
+export * from './types';
 export * from './utils';
